back-end/requirements.txt
fastapi[standard]>=0.128.0
psycopg[binary]>=3.3.2
python-dotenv
